Authors and users edit

+++ src/app/authors/authors.component.ts
@@ -3,6 +3,7 @@ import {LoggingService} from '../logging.service';
 import Tabulator from 'tabulator-tables';
 import {ActivatedRoute, Router} from '@angular/router';
 import {AuthorsService} from './authors.service';
+import {Author} from "./author.model";
 
 @Component({
     selector: 'app-users',
@@ -12,6 +13,10 @@ import {AuthorsService} from './authors.service';
 })
 export class AuthorsComponent implements OnInit {
 
+    private authors: Author[] = [];
+    private table;
+    searchTerm: string = "";
+
     constructor(private router: Router,
                 private route: ActivatedRoute,
                 private loggingService: LoggingService,
@@ -20,8 +25,6 @@ export class AuthorsComponent implements OnInit {
     }
 
     ngOnInit() {
-        var tabledata = this.authorsServer.getAuthors();
-
         var editIcon = function (cell, formatterParams, onRendered) {
             return '<i class=\'fas fa-pencil-alt\'></i>';
         };
@@ -30,14 +33,30 @@ export class AuthorsComponent implements OnInit {
             this.router.navigate(['edit-authors/' + id]);
         };
 
-        var table = new Tabulator('#authors-table', {
-            data: tabledata,
+        this.table = new Tabulator('#authors-table', {
+            data: this.authors,
             height: 'calc(100vh - 250px)',
             layout: 'fitColumns',
             addRowPos: 'top',
-            pagination: 'local',
-            paginationSize: 6,
-            paginationSizeSelector: [3, 6, 8, 10],
+            ajaxFiltering: true,
+            pagination: "remote",
+            ajaxConfig: "POST",
+            paginationDataSent: { // override the sent params for pagination
+                "page": "pageNumber",
+                "size": "pageSize"
+            },
+            paginationDataReceived: { // override the received pagination default properties
+                "last_page": "totalPages",
+                "current_page": "number",
+                "data": "content"
+            },
+            paginationSizeSelector: [5, 10, 15, 20, 50, 100, 500, 1000],
+            ajaxURL: "http://localhost:8080/api/v1/authors/filter",
+            ajaxURLGenerator: (url, config, params) => {
+                params.searchTerm = this.searchTerm;
+                return url + "?params=" + encodeURI(JSON.stringify(params));
+            },
+            paginationSize: 20,
             initialSort: [
                 {column: 'name', dir: 'asc'},
             ],
@@ -46,34 +65,29 @@ export class AuthorsComponent implements OnInit {
                 {title: 'Id', field: 'id', width: 60},
                 {title: 'First name', field: 'firstName'},
                 {title: 'Last name', field: 'lastName'},
-                {
-                    title: 'Delete',
-                    formatter: 'buttonCross',
-                    align: 'center',
-                    width: 100,
-                    headerSort: false,
-                    cellClick: function (e, cell) {
-                        cell.getRow().delete();
-                    }
-                },
+                {title: 'Delete', formatter: 'buttonCross', align: 'center', width: 100, headerSort: false},
             ],
             rowDblClick: (e, row) => {
                 this.router.navigate([`/edit-authors/${row.getData().id}`]);
             },
             cellClick: (e, cell) => {
                 if (cell.getColumn().getDefinition().title == "Edit") {
-                    this.router.navigate([`/edit-authors/${cell.getRow().getData().id}`]);
+                    this.router.navigate([`/edit-authors/${cell.getData().id}`]);
+                } else if(cell.getColumn().getDefinition().title == "Delete"){
+                    this.authorsServer.deleteAuthor(cell.getData().id);
+                    cell.getRow().delete();
                 }
             }
         });
+        this.table.setData();
     }
 
-    onNewUser() {
+    onNewAuthor() {
         this.router.navigate(['edit-authors/0']);
     }
 
     onSearch() {
-
+        this.table.setData();
     }
 
 }
